Initial commit
- added comments and did some clean up work
