+++ imports/api/s3Storage.js
@@ -1,18 +1,30 @@
+/*
+ * Amazon S3 related API. Used to upload, download, and maintain files on S3 servers.
+ */
+
+//Read private data for setting up AWS
 let rawCredentials = JSON.parse(Assets.getText("AWSConfig.json"));
 
-//Slingshot S3 Upload
-
+//Credentials for AWS
 let credentials = {
     accessKeyId: rawCredentials.accessKeyId,
     secretAccessKey: rawCredentials.secretAccessKey,
     region: rawCredentials.region
 }
 
+//Restrict type of files that can be uploaded
 Slingshot.fileRestrictions("userImageUploads", {
-    allowedFileTypes: ["image/png", "image/jpeg"],
-    maxSize: 25 * 1024 * 1024 // 10 MB (use null for unlimited).
+    allowedFileTypes: ["image/png", "image/jpeg"], // Limit file upload type to png / jpeg files
+    maxSize: 25 * 1024 * 1024 //Limit file upload size to 10MB.
 });
 
+/**
+ * Function to upload user images. Uses a library called slingshot.
+ *
+ * Creates an object that can be used to send user images to S3. See image widget main file for examples.
+ *
+ * Only accepts valid files and returns a download URL
+ */
 Slingshot.createDirective("userImageUploads", Slingshot.S3Storage, {
     cdn: 'https://s3.ap-northeast-2.amazonaws.com/ryeboard/',
     bucket: "ryeboard",
@@ -35,9 +47,9 @@ Slingshot.createDirective("userImageUploads", Slingshot.S3Storage, {
 });
 
 //Raw S3 API
-
 let S3 = require('aws-sdk/clients/s3');
 
+//Create correct credentials to start using raw S3 API
 let s3 = new S3({
     apiVersion: '2006-03-01',
     region: credentials.region,
@@ -48,6 +60,12 @@ let s3 = new S3({
 });
 
 Meteor.methods({
+
+    /**
+     * Boards Initialize
+     *
+     * Cleans up S3 storage of images not being used.
+     */
     'boards.initialize'(){
         if (this.userId) {
 
@@ -77,10 +95,10 @@ Meteor.methods({
                                 if (err) console.log(err, err.stack); // an error occurred
                                 else     console.log(data);           // successful response
                             });
-                            console.log('deleting:', "https://ryeboard.s3.ap-northeast-2.amazonaws.com/ryeboard/userData/" + key);
                         }
                     });
                 });
+
             });
 
         }
