+++ imports/helper/fillerText.js
@@ -1,3 +1,11 @@
+/**
+ * Filler
+ *
+ * Returns latin filler text used for fast prototyping
+ *
+ * @param {String} length - 'short', 'long', null
+ * @returns {String} - appropriate length of text
+ */
 export default function filler(length){
     if(length==="short"){
         return "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor."
