+++ imports/helper/isUrlImage.js
@@ -1,4 +1,11 @@
-export default function testImage(url, callback, timeoutT) {
+/**
+ * Checks to see if URL is an image
+ *
+ * @param {String} url - url that is being tested
+ * @param {Function} callback - function to call after checking
+ * @param {Integer} timeoutT - how long it should wait for seconds
+ */
+export default function testIfImage(url, callback, timeoutT) {
     let timeout = timeoutT || 5000;
     let timer, img = new Image();
     img.onerror = img.onabort = function () {
