+++ server/main.js
@@ -3,6 +3,7 @@ import '../imports/api/boards.js';
 import '../imports/api/s3Storage.js';
 import '/imports/startup/server/init';
 
+//Fix using images from other sources
 Meteor.startup(() => {
     WebApp.rawConnectHandlers.use(function(req, res, next) {
         res.setHeader("Access-Control-Allow-Origin", "*");
